Move type and constants to constants.ts
